minor fixes

+++ Volume/header/RayMarch.hpp
@@ -17,10 +17,10 @@
 
 void render(const int img_w, const int img_h, std::shared_ptr<Camera> camera, const lux::SField& sfield, const lux::CField& cfield);
 lux::Color marchRays(lux::Vector pos, lux::Vector dir, const lux::SField& field, const lux::CField& c);
-double marchRaysDSM(lux::Vector pos, lux::Vector lightPos, const lux::SField& density);
+float marchRaysDSM(lux::Vector pos, lux::Vector lightPos, const lux::SField& density);
 
 // with lights
 // removed const - fix it!
 void render(const int img_w, const int img_h, std::shared_ptr<Camera> camera, lux::SField& g, std::vector<std::shared_ptr<lux::Light>>& lights);
 lux::Color marchRays(lux::Vector pos, lux::Vector dir, const lux::SField& g, const std::vector<std::shared_ptr<lux::Light>>& lights);
-double marchToLight(lux::Vector pos, lux::Vector lightPos, const lux::SField& g);
+float marchToLight(lux::Vector pos, lux::Vector lightPos, const lux::SField& g);
